feat(docs): refactor DocsSidebar for improved layout and responsiveness
